requirements.txt
discord.py==2.5.2
youtube_dl==2021.12.17
PyNaCl==1.5.0
python-dotenv==1.1.1